@asubiotto 2023-02-16
Introducing bugs that I then fix is what I call job security.